backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hume>=0.11.3",
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "python-multipart>=0.0.6",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "ecdsa==0.19.1",
    "h11==0.16.0",
    "idna==3.10",
    "pyasn1==0.6.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "python-jose==3.5.0",
    "requests==2.32.5",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.47.3",
    "typing-inspection==0.4.1",
    "typing-extensions==4.15.0",
    "urllib3==2.5.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.8",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
